Fix execution order in LUOobCommand causing wrongly setup node list
In commit bfceedbe a check was added to put the master at the end or
skip it completely. While this functionality works, it was done at the
wrong point because node_names was already processed to a node list
without this checks. The Exec method if the LU just used that node list.

To overcome this situation we correct the node_names before the node
list is generated.

Signed-off-by: René Nussbaumer <rn@google.com>
Reviewed-by: Iustin Pop <iustin@google.com>