Fix dist archive generation
As reported in issue 140, the timestamps of the generated files can be
older than the commit timestamp, leading to needless rebuilds of man
pages and such. Furthermore, with the man pages not built, it
currently fails, as there is no dependency on them.

Therefore, we change the ‘dist’ rule so that it forces a
maintainer-clean and then explicitly rebuilds the needed targets, so
that we have up-to-date timestamps.

Last, we also ship the HTML versions of the man pages.

Signed-off-by: Iustin Pop <iustin@google.com>
Reviewed-by: Michael Hanselmann <hansmi@google.com>